extracts types related to variable handling to its own files
