+++ Cecilifier.Core/AST/EnumDeclarationVisitor.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Cecilifier.Core.Extensions;
 using Cecilifier.Core.Misc;
-using Microsoft.CodeAnalysis;
+using Cecilifier.Core.Variables;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
@@ -33,7 +33,7 @@ namespace Cecilifier.Core.AST
                 throw new NullReferenceException($"Something really bad happened. Roslyn failed to resolve the symbol for the enum {node.Identifier.Text}");
             
             var parentName = enumSymbol.ContainingSymbol.FullyQualifiedName();
-            using (Context.DefinitionVariables.WithCurrent(parentName, node.Identifier.ValueText, MemberKind.Type,enumType))
+            using (Context.DefinitionVariables.WithCurrent(parentName, node.Identifier.ValueText, VariableMemberKind.Type,enumType))
             {
                 //.class private auto ansi MyEnum
                 var fieldVar = Context.Naming.LocalVariable(node, "valueField");
@@ -52,7 +52,7 @@ namespace Cecilifier.Core.AST
             // Adds a field like:
             // .field public static literal valuetype xxx.MyEnum Second = int32(1)
             var enumMemberValue = _memberCollector[node];
-            var enumVarDef = Context.DefinitionVariables.GetLastOf(MemberKind.Type);
+            var enumVarDef = Context.DefinitionVariables.GetLastOf(VariableMemberKind.Type);
 
             var fieldVar = Context.Naming.LocalVariable(node, $"em_{enumVarDef.MemberName}");
             var exp = CecilDefinitionsFactory.Field(enumVarDef.VariableName, fieldVar, node.Identifier.ValueText, enumVarDef.VariableName,
