+++ Cecilifier.Core/AST/IfStatementVisitor.cs
@@ -1,4 +1,5 @@
 ﻿using Cecilifier.Core.Mappings;
+using Cecilifier.Core.Variables;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Mono.Cecil.Cil;
 
@@ -47,7 +48,7 @@ namespace Cecilifier.Core.AST
             }
 
             WriteCecilExpression(Context, $"{_ilVar}.Append({elseEndTargetVarName});");
-            WriteCecilExpression(Context, $"{Context.DefinitionVariables.GetLastOf(MemberKind.Method).VariableName}.Body.OptimizeMacros();");
+            WriteCecilExpression(Context, $"{Context.DefinitionVariables.GetLastOf(VariableMemberKind.Method).VariableName}.Body.OptimizeMacros();");
             Context.WriteComment($" end if ({node.Condition})");
         }
     }
