+++ Cecilifier.Core/Variables/VariableMemberKind.cs
@@ -1,0 +1,12 @@
+namespace Cecilifier.Core.Variables;
+
+public enum VariableMemberKind
+{
+    Type,
+    TypeParameter,
+    Field,
+    Method,
+    Parameter,
+    LocalVariable,
+    ModuleReference
+}
